Extend and refactor `IcebergDataset` logging to indicate volume of filepaths accumulated
